（杂项修改）拆分配置文件、升级主题
